pyproject.toml
[project]
name = "orca-ztbus-python-processor"
version = "0.1.0"
description = ""
authors = [
    {name = "Predixus",email = "contact@predixus.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary[pool] (>=2.9.10,<3.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "orca-python >=0.5.0"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
types-psycopg2 = "^2.9.21.20250516"
tqdm = "^4.67.1"
poethepoet = "^0.35.0"
numpy = "^2.3.1"
mypy = "^1.16.1"
ruff = "^0.12.0"

[tool.poe.tasks]
_lint_fix = "ruff check ./*.py --fix "
_type = "mypy ./*.py --strict --warn-unused-ignores --warn-redundant-casts "
_format = "ruff format ./*.py"

format = ["_format", "_lint_fix"]
lint = ["_format", "_lint_fix", "_type"]



